feat(chat-reaction): add reaction toggle
